# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CAPEF Digital Enrollment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List all app users (admin only)
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, supervisor, agent]
        - in: query
          name: regionId
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create/invite a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [auth]
      summary: Get a single user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [auth]
      summary: Update a user's role or region (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete a user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Reference Data ───────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Aggregated statistics for the dashboard
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: activity
          schema:
            type: string
        - in: query
          name: regionId
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/recent:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent enrollment activity feed
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberSummary"

  # ─── Uploads ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Auth / Users ────────────────────────────────────────────────────────
    ZoneAssignment:
      type: object
      required: [regionId]
      properties:
        regionId:
          type: integer
        regionName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        arrondissementId:
          type: ["integer", "null"]
        arrondissementName:
          type: ["string", "null"]
    AppUser:
      type: object
      required: [id, clerkUserId, email, name, role, status, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        regionId:
          type: ["integer", "null"]
        regionName:
          type: ["string", "null"]
        cniNumber:
          type: ["string", "null"]
        cniPhotoUrl:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, banned]
        assignedZones:
          type: array
          items:
            $ref: "#/components/schemas/ZoneAssignment"
        createdAt:
          type: string
...
    AppUserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        regionId:
          type: ["integer", "null"]
        cniNumber:
          type: ["string", "null"]
        cniPhotoUrl:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, banned]
        assignedZones:
          type: array
          items:
            $ref: "#/components/schemas/ZoneAssignment"
    AppUserUpdate:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        regionId:
          type: ["integer", "null"]
        cniNumber:
          type: ["string", "null"]
        cniPhotoUrl:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, banned]
        assignedZones:
          type: array
          items:
            $ref: "#/components/schemas/ZoneAssignment"

    # ─── Reference Data ──────────────────────────────────────────────────────
...
    MemberSummary:
      type: object
      required: [id, memberNumber, memberType, category, createdAt, status]
      properties:
        id:
          type: integer
        memberNumber:
          type: string
        memberType:
          type: string
          enum: [physique, morale]
        category:
          type: string
          enum: [agriculteur, pecheur, eleveur, forestier, artisan]
        displayName:
          type: ["string", "null"]
        regionName:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        badgeUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [incomplet, en_attente, valide, desactive, bloque]
        createdAt:
          type: string
...
    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer
    RegionCount:
      type: object
      required: [regionName, count]
      properties:
        regionName:
          type: string
        count:
          type: integer
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
    DashboardStats:
      type: object
      required: [totalMembers, totalPhysique, totalMorale, organisationsRepresenteesParFemmes, byCategory, byRegion, recentWeekCount]
      properties:
        totalMembers:
          type: integer
        totalPhysique:
          type: integer
        totalMorale:
          type: integer
        organisationsRepresenteesParFemmes:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        byRegion:
          type: array
          items:
            $ref: "#/components/schemas/RegionCount"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        recentWeekCount:
          type: integer

    # ─── Uploads ─────────────────────────────────────────────────────────────
...