# lib/api-zod/src/generated/types/appUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CAPEF Digital Enrollment API
 * OpenAPI spec version: 0.1.0
 */
import type { AppUserRole } from './appUserRole';
import type { AppUserStatus } from './appUserStatus';
import type { ZoneAssignment } from './zoneAssignment';

export interface AppUser {
  id: number;
  clerkUserId: string;
  email: string;
  name: string;
  role: AppUserRole;
  /** @nullable */
  regionId?: number | null;
  /** @nullable */
  regionName?: string | null;
  /** @nullable */
  cniNumber?: string | null;
  /** @nullable */
  cniPhotoUrl?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  status: AppUserStatus;
  assignedZones?: ZoneAssignment[];
  createdAt: string;
}