# lib/api-zod/src/generated/types/appUserInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CAPEF Digital Enrollment API
 * OpenAPI spec version: 0.1.0
 */
import type { AppUserInputRole } from './appUserInputRole';
import type { AppUserInputStatus } from './appUserInputStatus';
import type { ZoneAssignment } from './zoneAssignment';

export interface AppUserInput {
  email: string;
  name: string;
  role: AppUserInputRole;
  /** @nullable */
  regionId?: number | null;
  /** @nullable */
  cniNumber?: string | null;
  /** @nullable */
  cniPhotoUrl?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  status?: AppUserInputStatus;
  assignedZones?: ZoneAssignment[];
}