# lib/api-zod/src/generated/types/appUserUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CAPEF Digital Enrollment API
 * OpenAPI spec version: 0.1.0
 */
import type { AppUserUpdateRole } from './appUserUpdateRole';
import type { AppUserUpdateStatus } from './appUserUpdateStatus';
import type { ZoneAssignment } from './zoneAssignment';

export interface AppUserUpdate {
  email?: string;
  name?: string;
  role?: AppUserUpdateRole;
  /** @nullable */
  regionId?: number | null;
  /** @nullable */
  cniNumber?: string | null;
  /** @nullable */
  cniPhotoUrl?: string | null;
  /** @nullable */
  profilePhotoUrl?: string | null;
  status?: AppUserUpdateStatus;
  assignedZones?: ZoneAssignment[];
}