# lib/api-zod/src/generated/types/zoneAssignment.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CAPEF Digital Enrollment API
 * OpenAPI spec version: 0.1.0
 */

export interface ZoneAssignment {
  regionId: number;
  /** @nullable */
  regionName?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  arrondissementId?: number | null;
  /** @nullable */
  arrondissementName?: string | null;
}